[NOTIFICATIONS] Add Notifications API's

+++ platform/microservices/trust-certification-system-core/src/main/java/com/dreamsoftware/tcs/model/events/OnNewIssueCertificateRequestEvent.java
@@ -34,5 +34,5 @@ public class OnNewIssueCertificateRequestEvent {
     /**
      * Qualification
      */
-    private Float qualification;
+    private Long qualification;
 }
